# nsflow/backend/api/v1/agent_flows.py
# nsflow/backend/api/v1/agent_flows.py

import os
import logging
from pathlib import Path
from fastapi import APIRouter, HTTPException
from nsflow.backend.utils.agent_network_utils import AgentNetworkUtils

logging.basicConfig(level=logging.INFO)

router = APIRouter(prefix="/api/v1")
agent_utils = AgentNetworkUtils()  # Instantiate utility class


@router.get("/networks/")
def get_networks():
    """Returns a list of available agent networks."""
    return agent_utils.list_available_networks()


@router.get("/network/{network_name}", responses={200: {"description": "Agent Network found"}, 
                                                  404: {"description": "Agent Network not found"}})
def get_agent_network(network_name: str):
    """Retrieves the network structure for a given agent network."""
    file_path = Path(os.path.join(agent_utils.registry_dir, f"{network_name}.hocon"))
    logging.info("file_path: %s", file_path)
    if not file_path.exists():
        raise HTTPException(status_code=404, detail=f"Network name '{network_name}' not found.")
    return agent_utils.parse_agent_network(file_path)


@router.get("/connectivity/{network_name}", responses={200: {"description": "Connectivity Info"}, 
                                                       404: {"description": "HOCON file not found"}})
def get_connectivity_info(network_name: str):
    """Retrieves connectivity details from an HOCON network configuration file."""
    file_path = Path(os.path.join(agent_utils.registry_dir, f"{network_name}.hocon"))
    if not file_path.exists():
        raise HTTPException(status_code=404, detail=f"HOCON file '{network_name}.hocon' not found.")
    return agent_utils.extract_connectivity_info(file_path)