# tests/fixtures/openapi.json
{"openapi":"3.0.2","info":{"title":"FastAPI","version":"0.1.0"},"paths":{"/api/v1/ws/connectivity/{agent_name}":{"get":{"summary":"Get Agent Connectivity","description":"Fetch agent connectivity details.","operationId":"get_agent_connectivity_api_v1_ws_connectivity__agent_name__get","parameters":[{"name":"agent_name","in":"path","required":true,"schema":{"type":"string","title":"Agent Name"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/ws/function/{agent_name}":{"get":{"summary":"Get Agent Function","description":"Fetch agent function details.","operationId":"get_agent_function_api_v1_ws_function__agent_name__get","parameters":[{"name":"agent_name","in":"path","required":true,"schema":{"type":"string","title":"Agent Name"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/networks/":{"get":{"summary":"Get Networks","description":"Returns a list of available agent networks.","operationId":"get_networks_api_v1_networks__get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/api/v1/network/{network_name}":{"get":{"summary":"Get Agent Network","description":"Retrieves the network structure for a given agent network.","operationId":"get_agent_network_api_v1_network__network_name__get","parameters":[{"name":"network_name","in":"path","required":true,"schema":{"type":"string","title":"Network Name"}}],"responses":{"200":{"description":"Agent Network found","content":{"application/json":{"schema":{}}}},"404":{"description":"Agent Network not found"},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/connectivity/{network_name}":{"get":{"summary":"Get Connectivity Info","description":"Retrieves connectivity details from an HOCON network configuration file.","operationId":"get_connectivity_info_api_v1_connectivity__network_name__get","parameters":[{"name":"network_name","in":"path","required":true,"schema":{"type":"string","title":"Network Name"}}],"responses":{"200":{"description":"Connectivity Info","content":{"application/json":{"schema":{}}}},"404":{"description":"HOCON file not found"},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/export/notebook/{agent_network}":{"get":{"summary":"Export Notebook","description":"Endpoint to generate and return a downloadable Jupyter Notebook for an agent network.","operationId":"export_notebook_api_v1_export_notebook__agent_network__get","parameters":[{"name":"agent_network","in":"path","required":true,"schema":{"type":"string","title":"Agent Network"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/export/agent_network/{agent_network}":{"get":{"summary":"Export Agent Network","description":"Endpoint to download the HOCON file of the selected agent network.","operationId":"export_agent_network_api_v1_export_agent_network__agent_network__get","parameters":[{"name":"agent_network","in":"path","required":true,"schema":{"type":"string","title":"Agent Network"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"404":{"description":"Agent network not found"},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}}},"components":{"schemas":{"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}}}}